requirements.txt
torchvision==0.3.0
torch==1.1.0
opencv_contrib_python==3.4.3.18
albumentations==0.1.12
numpy==1.16.5
Pillow==6.1.0
